Maak aparte maven module voor commandline tool
basis documentatie voor command line tool
en stop met packaging van de oude bgt commandline tool
