+++ brmo-commandline/src/main/resources/commandline-example.properties
@@ -1,5 +1,5 @@
 # Copyright (C) 2016 B3Partners B.V.
-# voorbeeld commadline voor de brmo commandline tool
+# voorbeeld commandline database configuratie voor de brmo commandline tool
 
 # database type
 #dbtype=oracle|postgis|jtds-sqlserver
@@ -8,16 +8,16 @@ dbtype=postgis
 # rsgb database gegevens
 rsgb.url=jdbc:postgresql://localhost:5433/itest_rsgb
 rsgb.user=rsgb
-rsgb.passwd=rsgb
+rsgb.password=rsgb
 
 # staging database gegevens
 staging.url=jdbc:postgresql://localhost:5433/itest_staging
 staging.user=rsgb
-staging.passwd=rsgb
+staging.password=rsgb
 # optionele staging db connectie properties in de vorm propertyName=property;propertyName2=property;
 #staging.options=""
 
 # rsgbbgt database gegevens
 rsgbbgt.url=jdbc:postgresql://localhost:5433/itest_rsgbbgt
 rsgbbgt.user=rsgb
-rsgbbgt.passwd=rsgb
+rsgbbgt.password=rsgb
