+++ brmo-commandline/src/main/resources/log4j.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE log4j:configuration PUBLIC "-//APACHE//DTD LOG4J 1.2//EN" "log4j.dtd">
+<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
+
+    <!-- deze appender schrijft naar de console -->
+    <appender name="consoleAppender" class="org.apache.log4j.ConsoleAppender">
+        <param name="Threshold" value="debug" />
+        <layout class="org.apache.log4j.EnhancedPatternLayout">
+            <!-- voor debug met regelnummers uit de broncode:
+            <param name="ConversionPattern" value="%5p %d{HH:mm:ss} (%C{1.}#%M:%L) - %m%n" / -->
+            <param name="ConversionPattern" value="%5p %d{HH:mm:ss.s} - %m%n" />
+        </layout>
+    </appender>
+
+    <!-- deze appender schrijft naar de genoemde logfile welke tot maximaal 5x wordt geroteerd bij 5MB -->
+    <appender name="fileAppender" class="org.apache.log4j.RollingFileAppender">
+        <param name="File" value="brmo-commandline.log"/>
+        <param name="Threshold" value="debug" />
+        <param name="Append" value="true" />
+        <param name="maxBackupIndex" value="5" />
+        <param name="maxFileSize" value="5MB" />
+        <layout class="org.apache.log4j.EnhancedPatternLayout">
+            <!-- voor debug met regelnummers uit de broncode:
+            <param name="ConversionPattern" value="%5p %d{HH:mm:ss} (%C{1.}#%M:%L) - %m%n" / -->
+            <param name="ConversionPattern" value="%5p %d{HH:mm:sss} - %m%n" />
+        </layout>
+    </appender>
+
+    <logger name="nl.b3p.brmo">
+        <!-- mogelijke values voor level zijn (in volgorde van verbositeit) trace,debug,info,warn,error,fatal -->
+        <level value="info" />
+    </logger>
+    <logger name="org.geotools">
+        <level value="info" />
+    </logger>
+    <logger name="hsqldb.db">
+        <level value="warn" />
+    </logger>
+    <root>
+        <level value="info" />
+        <appender-ref ref="fileAppender" />
+    </root>
+</log4j:configuration>
