+++ brmo-loader/src/main/java/nl/b3p/brmo/loader/BrmoFramework.java
@@ -477,4 +477,12 @@ public class BrmoFramework {
             throw new BrmoException(ex);
         }
     }
+
+    public long getCountJob() throws BrmoException {
+        try {
+            return stagingProxy.getCountJob();
+        } catch (SQLException ex) {
+            throw new BrmoException(ex);
+        }
+    }
 }
