+++ pom.xml
@@ -13,6 +13,7 @@
         <module>brmo-loader</module>
         <module>brmo-taglib</module>
         <module>brmo-service</module>
+        <module>brmo-commandline</module>
         <module>brmo-proxyservice</module>
         <module>brmo-soap</module>
         <module>brmo-dist</module>
